Refactor the langevin generator to its own module.
